+++ examples/sparql.tsx
@@ -73,13 +73,14 @@ function SparqlExample() {
                 ]}
                 languages={[
                     {code: 'de', label: 'Deutsch'},
-                    {code: 'en', label: 'english'},
-                    {code: 'es', label: 'español'},
-                    {code: 'fr', label: 'français'},
-                    {code: 'ja', label: '日本語'},
+                    {code: 'en', label: 'English'},
+                    {code: 'es', label: 'Español'},
+                    {code: 'fr', label: 'Français'},
                     {code: 'hi', label: 'हिन्दी'},
+                    {code: 'it', label: 'Italiano'},
+                    {code: 'ja', label: '日本語'},
                     {code: 'pt', label: 'português'},
-                    {code: 'ru', label: 'русский'},
+                    {code: 'ru', label: 'Русский'},
                     {code: 'zh', label: '汉语'},
                 ]}
             />